map_token_conflict fix

+++ contracts/child/ChildChainManager/IChildChainManager.sol
@@ -1,8 +1,8 @@
 pragma solidity 0.6.6;
 
 interface IChildChainManager {
-    event TokenMapped(address indexed rootToken, address indexed childToken, uint256 indexed chainId);
+    event TokenMapped(address indexed rootToken, address indexed childToken, uint64 indexed chainId);
 
-    function mapToken(address rootToken, address childToken, uint256 chainId) external;
-    function cleanMapToken(address rootToken, address childToken, uint256 chainId) external;
+    function mapToken(address rootToken, address childToken, uint64 chainId) external;
+    function cleanMapToken(address rootToken, address childToken, uint64 chainId) external;
 }
